Added "keylow" and "keyhigh" option to axes, which generates key events when the axis is moved.
One can specify scancodes for negative and positive axis movement. When the axis is moved out of the deadzone,
keydown events are once generated, and keyrelease events are generated when the axis returns to 0.
One can map a pad to the four arrow keys to play games that do not support a joystick.