Key autorepeat added for axes with mode=relative
If axis is set to relative and keylow and keyhigh are assigned, autorepeat for these keys is emulated with a speed relative to the
value of the axis (keydown and keyup are generated frequently). A factor for this autorepeat speed can be assigned by
"axis=<factor>key".
If mode is "accelerated", the axis behaves like a button: only one keydown and one keyup, X does autorepeat for the key.