Readded own keymap for Joystick keys. This changes meaning of key, keylow and keyhigh values in xorg.conf from scancodes to keysyms!
New xorg-server 1.4 fixes a lot of keyboard things. This makes it necessary to
provide an own mapping rather than using the mapping of the current keyboard. A
keymap is built dynamically from the options provided. The options now need to
provide a keysym for each button instead of the scancodes of another keyboard
attached keyboard.