Always add second input device for keys only
Having one device for both mouse and keyboard events breaks with MPX and custom keymaps, because the keymap is
copied to the core pointer instead of the core keyboard.

Easiest fix for that is to have two separate input devices. One single instance of the driver will now create
a second input device that has only keyboard capabilities. It shares all the config with the master and will
be removed when the master device is removed.

The creation of this second device is a little hack-ish. I hope things are done right.