Free pInfo->private only once to fix server crash on unplug (bug #35391)
The hotplugged keyboard device and the main device share the same pInfo->private data
and the same jstkCoreUnInit, so the data is freed twice. Furthermore, since the keyboard
device will delete itself, we must not delete it from within the deletion of the main device.

Freeing pInfo->private is done by the main device, hopefully done independently of the order
in which the two devices will be removed by the server.

Should fix bug #35391