Copy pInfo->driver to option list to fix hotplugging of keyboard device
udev does copy the Driver line to the list of options, but when manually
specifying the driver in xorg.conf, the option "Driver" is unset. Because we
do hotplug a sub-device from within the core device, we need the "Driver"
option to be present in the list.

This should fix archlinux bug #23577:
  https://bugs.archlinux.org/task/23577

Thanks to Malek for coming up with a fix.

Signed-off-by: Sascha Hlusiak <contact@saschahlusiak.de>