config: warning in default xorg.conf.d snippet: do not enable by default
see discussion: https://bugs.gentoo.org/show_bug.cgi?id=403003

distributions should not ship the module enabled by default, because a lot
of users have the joystick module installed without knowing, resulting in
confusion and bad user experience when trying to play games, the more
likely usecase when plugging in a joystick.

the default configuration only fits for a limited number of users.

it does not make much sense to automatically enable hotplugging of the
module without explicit user consent and reviewing of the configuration.

Signed-off-by: Sascha Hlusiak <saschahlusiak@arcor.de>